# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta es la API de productos para la tienda.",
        "title": "API Tienda Online",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/categories": {
            "get": {
                "description": "Obtiene todas las categorías desde la base de datos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Listar categorías",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Category"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Crea una nueva categoría",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Crear categoría",
                "parameters": [
                    {
                        "description": "Categoría",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Category"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/categories/{id}": {
            "delete": {
                "description": "Elimina una categoría por su ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Eliminar categoría",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la categoría",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "Retorna una lista de productos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Obtener todos los productos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Product"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo producto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Crear un producto",
                "parameters": [
                    {
                        "description": "Producto a crear",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Product"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "delete": {
                "description": "Elimina un producto por ID",
                "tags": [
                    "products"
                ],
                "summary": "Eliminar un producto",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del producto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Category": {
            "type": "object",
            "properties": {
                "activo": {
                    "type": "boolean"
                },
                "descripcion": {
                    "type": "string"
                },
                "fechaActualizacion": {
                    "type": "string"
                },
                "fechaCreacion": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "descripcion": {
                    "type": "string"
                },
                "destacado": {
                    "type": "boolean"
                },
                "fecha_actualizacion": {
                    "type": "string"
                },
                "fecha_creacion": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imagen": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "precio": {
                    "type": "string"
                },
                "precio_descuento": {
                    "type": "string"
                },
                "publicado": {
                    "type": "boolean"
                },
                "stock": {
                    "type": "integer"
                }
            }
        }
    }
}